requirements.txt
python-telegram-bot==20.8
ollama==0.3.0
python-dotenv==1.0.0